Move `run_dbt_and_capture_stdout` into the test case
